[FIX] mail: fix activity state is shared accross tabs
Since [1], the activity state, which is supposed to be shared
accross tab through a broadcast channel, isn't anymore.

This PR fixes the responsible typo.

[1]: #161286

X-original-commit: 914aad663d5e5e06d2646390254d987563286232

+++ addons/mail/static/src/core/web/activity_model_patch.js
@@ -75,6 +75,9 @@ patch(Activity.prototype, {
         return action;
     },
     remove({ broadcast = true } = {}) {
+        if (!this.exists()) {
+            return;
+        }
         this.delete();
         if (broadcast) {
             this.store.activityBroadcastChannel?.postMessage({
